Edicao no CRUD Mantener clientes

+++ vicarfood/src/main/java/com/br/vicarfood/vicarfood/repository/BairroRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BairroRepository extends JpaRepository<Bairro, Integer>{
+public interface BairroRepository extends JpaRepository<Bairro, Long>{
     
 }
